requirements.txt
PyQt6>=6.4.0
Pillow>=10.0.0
numpy>=1.24.0
opencv-python>=4.8.0
pytesseract>=0.3.10
